improving the product list style and separate it's code into more compenent
